+++ packages/templates/minimal/vite.config.ts
@@ -1,4 +1,4 @@
-import { aero } from 'aerobuilt/vite'
+import { aero } from '@aero-js/vite'
 import { defineConfig } from 'vite'
 
 export default defineConfig({